+++ dbt/adapters/synapse/__init__.py
@@ -1,6 +1,7 @@
 from dbt.adapters.base import AdapterPlugin
 
 from dbt.adapters.synapse.synapse_adapter import SynapseAdapter
+from dbt.adapters.synapse.synapse_column import SynapseColumn
 from dbt.adapters.synapse.synapse_connection_manager import SynapseConnectionManager
 from dbt.adapters.synapse.synapse_credentials import SynapseCredentials
 from dbt.include import synapse
@@ -12,4 +13,10 @@ Plugin = AdapterPlugin(
     dependencies=["fabric"],
 )
 
-__all__ = ["Plugin", "SynapseConnectionManager", "SynapseAdapter", "SynapseCredentials"]
+__all__ = [
+    "Plugin",
+    "SynapseConnectionManager",
+    "SynapseColumn",
+    "SynapseAdapter",
+    "SynapseCredentials",
+]